+++ grpc/middleware/correlationid/generator.go
@@ -1,6 +1,0 @@
-package correlationid
-
-type IdGenerator interface {
-	// Generate generates a unique string which can be used as a correlation ID.
-	Generate() string
-}