+++ grpc/middleware/correlationid/metadata_test.go
@@ -3,11 +3,12 @@ package correlationid_test
 import (
 	"testing"
 
+	"context"
+
 	"github.com/goph/fxt/grpc/middleware/correlationid"
-	"github.com/goph/fxt/grpc/middleware/correlationid/internal/mocks"
+	"github.com/goph/fxt/internal/correlationid/mocks"
 	"github.com/grpc-ecosystem/go-grpc-middleware/util/metautils"
 	"github.com/stretchr/testify/assert"
-	"golang.org/x/net/context"
 )
 
 func TestMetadataCarrier_GetCorrelationID(t *testing.T) {