+++ grpc/middleware/correlationid/server_interceptors.go
@@ -1,18 +1,20 @@
 package correlationid
 
 import (
+	"context"
+
+	"github.com/goph/fxt/internal/correlationid"
 	"github.com/grpc-ecosystem/go-grpc-middleware"
 	"github.com/grpc-ecosystem/go-grpc-middleware/tags"
-	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 )
 
 const TagCorrelationID = "correlationid"
 
 // UnaryServerInterceptor returns a new unary server interceptor for propagating correlation ID.
-func UnaryServerInterceptor(idGenerator IdGenerator, carrier Carrier) grpc.UnaryServerInterceptor {
+func UnaryServerInterceptor(generator correlationid.Generator, carrier correlationid.Carrier) grpc.UnaryServerInterceptor {
 	return func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
-		ctx = serverCorrelationID(idGenerator, carrier, ctx)
+		ctx = serverCorrelationID(generator, carrier, ctx)
 
 		resp, err := handler(ctx, req)
 
@@ -21,9 +23,9 @@ func UnaryServerInterceptor(idGenerator IdGenerator, carrier Carrier) grpc.Unary
 }
 
 // StreamServerInterceptor returns a new streaming server interceptor for propagating correlation ID.
-func StreamServerInterceptor(idGenerator IdGenerator, carrier Carrier) grpc.StreamServerInterceptor {
+func StreamServerInterceptor(generator correlationid.Generator, carrier correlationid.Carrier) grpc.StreamServerInterceptor {
 	return func(srv interface{}, stream grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
-		ctx := serverCorrelationID(idGenerator, carrier, stream.Context())
+		ctx := serverCorrelationID(generator, carrier, stream.Context())
 
 		wrappedStream := grpc_middleware.WrapServerStream(stream)
 		wrappedStream.WrappedContext = ctx
@@ -34,11 +36,11 @@ func StreamServerInterceptor(idGenerator IdGenerator, carrier Carrier) grpc.Stre
 	}
 }
 
-func serverCorrelationID(idGenerator IdGenerator, carrier Carrier, ctx context.Context) context.Context {
+func serverCorrelationID(generator correlationid.Generator, carrier correlationid.Carrier, ctx context.Context) context.Context {
 	correlationID, ok := carrier.GetCorrelationID(ctx)
 	if !ok {
 		if correlationID == "" {
-			correlationID = idGenerator.Generate()
+			correlationID = generator.Generate()
 		}
 
 		ctx = carrier.SetCorrelationID(ctx, correlationID)