+++ internal/correlationid/interfaces.go
@@ -1,7 +1,11 @@
 package correlationid
 
-import "golang.org/x/net/context"
+import "context"
 
+//go:generate mockery -name=Carrier
+//go:generate mockery -name=Generator
+
+// Carrier retrieves and packs a correlation ID from/into a context.
 type Carrier interface {
 	// GetCorrelationID returns an existing correlation ID.
 	// If the second argument is false, the interceptor attempts to set the correlation ID.
@@ -10,4 +14,10 @@ type Carrier interface {
 
 	// SetCorrelationID sets a correlation ID in the context and returns the new context (if any).
 	SetCorrelationID(ctx context.Context, correlationID string) context.Context
+}
+
+// Generator generates a correlation ID.
+type Generator interface {
+	// Generate generates a unique string which can be used as a correlation ID.
+	Generate() string
 }